+++ src/test/kotlin/com/github/anjoismysign/bloblibide/MyPluginTest.kt
@@ -1,12 +1,10 @@
 package com.github.anjoismysign.bloblibide
 
 import com.intellij.ide.highlighter.XmlFileType
-import com.intellij.openapi.components.service
 import com.intellij.psi.xml.XmlFile
 import com.intellij.testFramework.TestDataPath
 import com.intellij.testFramework.fixtures.BasePlatformTestCase
 import com.intellij.util.PsiErrorElementUtil
-import com.github.anjoismysign.bloblibide.services.MyProjectService
 
 @TestDataPath("\$CONTENT_ROOT/src/test/testData")
 class MyPluginTest : BasePlatformTestCase() {
@@ -30,9 +28,9 @@ class MyPluginTest : BasePlatformTestCase() {
     }
 
     fun testProjectService() {
-        val projectService = project.service<MyProjectService>()
-
-        assertEquals(4, projectService.getRandomNumber())
+//        val projectService = project.service<MyProjectService>()
+//
+//        assertEquals(4, projectService.getRandomNumber())
     }
 
     override fun getTestDataPath() = "src/test/testData/rename"