+++ CHANGELOG.md
@@ -3,5 +3,7 @@
 # bloblib-ide Changelog
 
 ## [Unreleased]
+
 ### Added
-- Initial scaffold created from [IntelliJ Platform Plugin Template](https://github.com/JetBrains/intellij-platform-plugin-template)
+
+- BlobObject class generator